google_live_api_agent/gemini_python/requirements.txt
websockets>=12.0
opencv-python>=4.8.0
pyaudio>=0.1.13
pillow>=10.0.0
mss>=9.0.0
taskgroup>=0.2.2
exceptiongroup>=1.1.0
python-dotenv>=1.0.0
numpy>=1.24.0
sounddevice>=0.4.6